Add initial object tracking image.
The object tracking code (CppMT) is built locally and then copied into the
image. This means that we only need to install the binary version of OpenCV and
not all the headers and compilers that would be needed to build in situ, saving
around 50% of disk space for the image.